+++ lib/ngtcp2_bbr2.c
@@ -1284,10 +1284,10 @@ static void bbr_bound_cwnd_for_model(ngtcp2_bbr2_cc *bbr,
 }
 
 static void bbr_set_send_quantum(ngtcp2_bbr2_cc *bbr, ngtcp2_conn_stat *cstat) {
-  uint64_t send_quantum =
-      (uint64_t)(cstat->pacing_rate * (double)(bbr->min_rtt == UINT64_MAX
-                                                   ? NGTCP2_MILLISECONDS
-                                                   : bbr->min_rtt));
+  size_t send_quantum =
+      (size_t)(cstat->pacing_rate * (double)(bbr->min_rtt == UINT64_MAX
+                                                 ? NGTCP2_MILLISECONDS
+                                                 : bbr->min_rtt));
   (void)bbr;
 
   cstat->send_quantum = ngtcp2_min(send_quantum, 64 * 1024);
